examples: Generalize event loop (main function) into shared support module
The Android example will use the same event loop code, and only needs a
different "entrypoint" (an example with `crate-type = cdylib`).  No
functional changes have been made to easen review; those will be applied
in a separate PR.
